Migrate lib/tokens.js to TypeScript
