+++ lib/skyline/token.ts
@@ -5,6 +5,8 @@
 //
 // Thank you so much for your original work!
 
+import { Tag } from "../tag";
+
 export class Loc {
 	constructor(public line: number, public column: number) { }
 }
@@ -12,7 +14,7 @@ export class Loc {
 export class Token {
 	loc: { start: Loc, end: Loc };
 
-	constructor(public tag: string, loc: { start: Loc, end: Loc }) {
+	constructor(public tag: string | Tag, loc: { start: Loc, end: Loc }) {
 		this.tag = tag;
 		this.loc = loc;
 	}
